requirements.txt
# Bot Library
discord.py==1.7.1
git+https://github.com/Rapptz/discord-ext-menus.git
# ^ discord-ext-menus isn't on pypi, but it will natively be in dpy 2.0 from what I can tell.

# Database
aiosqlite>=0.17.0

# Graph & Image manipulation
matplotlib==3.4.2

# Web Server
fastapi>=0.65.1
uvicorn[standard]>=0.13.4
pydantic>=1.8.2

# Fancy Extras
tabulate>=0.8.9
colorama>=0.4.4
